refactor/examples: extract boilerplate code to separate module
Makes it easier to read examples as there's less noise.
